refactor: 메서드 이동
