+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -8,14 +8,9 @@ import java.net.Socket;
 import java.nio.charset.StandardCharsets;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.common.MimeType;
-import org.apache.coyote.http11.common.HttpMethod;
-import org.apache.coyote.http11.common.HttpStatus;
 import org.apache.coyote.http11.handler.HandlerMapping;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.HttpRequestParser;
-import org.apache.coyote.http11.response.HttpResponse;
-import org.apache.coyote.http11.response.ResponseEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -45,9 +40,7 @@ public class Http11Processor implements Runnable, Processor {
         ) {
             HttpRequest httpRequest = HttpRequestParser.extract(reader);
 
-            ResponseEntity responseEntity = HANDLER_MAPPING.extractResponseEntity(httpRequest);
-
-            String response = extractResponse(responseEntity);
+            String response = HANDLER_MAPPING.extractResponseEntity(httpRequest);
 
             outputStream.write(response.getBytes());
             outputStream.flush();
@@ -55,28 +48,6 @@ public class Http11Processor implements Runnable, Processor {
             log.error(e.getMessage(), e);
         }
     }
-
-    private String extractResponse(ResponseEntity responseEntity) {
-        if (responseEntity.hasSameHttpMethod(HttpMethod.GET)) {
-            ViewResolver viewResolver = new ViewResolver(responseEntity);
-            HttpResponse httpResponse = viewResolver.extractHttpResponse();
-
-            return httpResponse.extractResponse();
-        }
-        return redirect(responseEntity);
-    }
-
-    private String redirect(ResponseEntity responseEntity) {
-        HttpStatus httpStatus = responseEntity.getHttpStatus();
-
-        return new StringBuilder()
-                .append(String.format("HTTP/1.1 %s %s ", httpStatus.getStatusCode(), httpStatus.name())).append("\r\n")
-                .append(String.format("Content-Type: %s;charset=utf-8 ", MimeType.HTML)).append("\r\n")
-                .append(String.format("Location: %s.html", responseEntity.getPath())).append("\r\n")
-                .append("\r\n")
-                .toString();
-    }
-
 
 }
